Add string repeat function.
In this commit, a new function 'repeatStr' is added. This function receives two parameters 'n' (an integer) and 's' (a string) and returns the string 's' repeated 'n' times. Considering this function can be used in various scenarios where a string needs to be repeated, it is useful to have this functionality incorporated in one function.