Implement abbreviation function for names
Added a JavaScript function 'abbrevName' to convert a full name (strictly two words separated by space) into initials. This function splits the name into an array, grabs the first character from each word, capitalizes them, and finally, joins them with a dot in between. This is useful when we want to display user names in a shorter format.