Add setAlarm functionality
The 'setAlarm' function is now added in the 8kyu/set-alarm.js file. This function takes in two parameters 'employed' and 'vacation'. It returns 'true' if the person is employed and not on vacation as these are the circumstances under which you need to set an alarm. This functionality will help to determine the need of setting an alarm depending upon the status of employment and vacation.