Add Opposites Attract function
Implemented a new JavaScript function called `lovefunc` in `opposites-attract.js` within the 8kyu directory. This function determines if a pair is in love based on the number of petals each one has. If one person has a flower with an even number of petals and the other has a flower with an odd number of petals, they are considered to be in love.