Your commit message does not match the diff provided. Please see the correct one below:
Commit message:
Add digitize function to convert number to reversed array

Added a JavaScript function called `digitize` to `convert-number-to-reversed-array-of-digits.js` within the 8kyu directory. This function takes a non-negative number and returns its digits as an array in reverse order. This can be very useful when working with digit manipulation tasks.