"Add makeNegative function in JavaScript
Created a new JavaScript function called 'makeNegative' that returns a negative version of a given number. This function is a simple solution for the task where a positive number needs to be converted into a negative number. And in case if the number is already negative, it will return the same number."