"Added number function in bus.js
A new function number has been added in 'number-of-people-in-the-bus.js'. This function returns the number of people who are still on the bus after the last bus stop, regardless of the number of people getting on or off at each stop. This function could be useful in transportation planning and management where such statistics may be needed."