"Implement deleteNth function to limit occurrence of an element
A new function named 'deleteNth' has been added to 'delete-occurrences-of-an-element-if-it-occurs-more-than-n-times.js'. This function takes an array and a number as input, and generates a new array where each number of the original array occurs at most n times, without reordering. This may be useful in cases requiring limit on repetition of same elements in an array."