"Add 'removeSmallest' function to remove the minimum value
Created a new JavaScript file '7kyu/remove-the-minimum.js' containing a function called 'removeSmallest'. This function takes an array of numbers, finds the minimum value, and removes the first instance of this minimum value from the array. If the input array is empty, it will return an empty array. This change will help in tasks where it is necessary to remove the smallest value from an array."