"Add nth term series sum function in JavaScript
The new JS file 'sum-of-the-first-nth-term-of-series.js' is added. The function 'SeriesSum' in this file calculates the sum of a specific mathematical series up to the nth term. The output is rounded to two decimal places and returned as a string. This function will be useful in mathematical computations involving series."