"Added towerBuilder function in JavaScript
A new file 'build-tower.js' is added, introductory comments are written to explain the function. This function, towerBuilder, builds a pyramid of "*" characters with a positive integer input for number of floors. It returns the pyramid as an array/list of strings. Several test cases are also added to verify the function's correctness."