"Added Tic-Tac-Toe game state checker function.
The new file 'tic-tac-toe-checker.js' in the 5kyu directory contains a function that checks the current state of a Tic-Tac-Toe game given a 3x3 array input. It will return 1 if "X" won, 2 if "O" won, -1 if the game is not yet finished, and 0 if the game ends in a draw. This functionality is important for assessing the outcome of the game at any given point."